requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.1
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.4.2
future==0.18.2
html5lib==1.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
webencodings==0.5.1
xhtml2pdf==0.2.5